code/requirements.txt
numpy~=1.22.1
pygame~=2.1.2
pandas~=1.4.0
pyqtgraph~=0.12.3
sklearn~=0.0
scikit-learn~=1.0.2
Pillow~=9.0.0
brainflow~=4.9.0
matplotlib~=3.5.1
imblearn~=0.0
mne~=0.24.1
scipy~=1.7.3